Submit an array job with a specified set of task IDs

Given a bash script that specifies the --array sbatch option (that is, an array job), this function overwrites (temporarily if restore = TRUE) the script in place and resubmits (when submit = TRUE) the array with the specified task_ids. If this array was created with job_single, task_ids may be ommitted and failed tasks are automatically inferred. This function is intended to help re-run failed tasks of a large array job that was previously submitted.

array_submit(
  job_bash,
  task_ids = NULL,
  submit = FALSE,
  restore = TRUE,
  verbose = FALSE
)

Arguments

job_bash

A character(1) vector with the name of a bash script in the current working directory.

task_ids

An optional numeric vector specifying which (relative) task IDs to resubmit (e.g. c(1, 4, 6)). If NULL, the task IDs will be inferred by scraping the log file for the job ID for the array job as originally submitted, and using job_report() to pull failed task IDs

submit

A logical(1) vector determining whether to actually submit the tasks or not using qsub.

restore

A logical(1) vector determining whether to restore the script to the original state.

verbose

A logical(1) vector specifying whether to print details about how failed tasks were determined (applicable when task_ids is NULL).

Value

The path to job_bash.

Author

Leonardo Collado-Torres

Nicholas J. Eagles

Examples

## Choose a script name
job_name <- paste0("array_submit_example_", Sys.Date())

## Create an array job on the temporary directory
with_wd(tempdir(), {
    ## Create an array job script to use for this example
    job_single(
        name = job_name,
        create_shell = TRUE,
        task_num = 100
    )

    ## Now we can submit the job for a set of task IDs (or omit 'task_ids'
    ## to automatically grab those same failed task IDs)
    array_submit(
        job_bash = paste0(job_name, ".sh"),
        task_ids = c(1, 6, 8:20, 67),
        submit = FALSE
    )
})
#> 2023-10-06 16:28:50.973317 creating the logs directory at:  logs
#> 2023-10-06 16:28:50.974784 creating the shell file array_submit_example_2023-10-06.sh
#> To submit the job use: sbatch array_submit_example_2023-10-06.sh
#> [1] "array_submit_example_2023-10-06.sh"