Build a bash script that loops over variables and submits SLURM jobs

This function builds a bash script that loops over a set of variables with pre-specified values to create an internal bash script that then gets submitted as a SLURM job.

job_loop(
  loops,
  name,
  create_shell = FALSE,
  partition = "shared",
  memory = "10G",
  cores = 1L,
  email = "ALL",
  logdir = "logs",
  task_num = NULL,
  tc = 20
)

Arguments

loops

A named list where each of the elements are character vectors. The names of loops specify the variables used for the loops and the contents specify the options to loop through for each variable.

name

A character(1) vector with the name of the script. Any spaces will be replaced by underscores.

create_shell

A logical(1) vector specifying whether to create a shell file for the script.

partition

A character(1) vector with the name of the SLURM partition. Check how busy a given partition is by running squeue -p [partition].

memory

The amount of memory per core to request. You can check how much a current job is utilizing using the sstat command.

cores

The number of cores to request. Note that the total memory your job will request is cores multiplied by memory.

email

The email reporting option for the email address ("BEGIN", "END", "FAIL", or "ALL")

logdir

The directory for the log files relative to the current working directory.

task_num

The number of tasks for your job, which will make it into an array job. If NULL this is ignored.

tc

If task_num is specified, this option controls the number of concurrent tasks.

Value

A character vector with the script contents. If create_shell was specified then it also creates the actual script in the current working directory.

Author

Leonardo Collado-Torres

Nicholas J. Eagles

Examples

job_loop(
    loops = list(region = c("DLPFC", "HIPPO"), feature = c("gene", "exon", "tx", "jxn")),
    name = "bsp2_test"
)
#> #!/bin/bash
#> 
#> ## Usage:
#> # sh bsp2_test.sh
#> 
#> ## Create the logs directory
#> mkdir -p logs
#> 
#> for region in DLPFC HIPPO; do
#>     for feature in gene exon tx jxn; do
#> 
#>     ## Internal script name
#>     SHORT="bsp2_test_${region}_${feature}"
#> 
#>     # Construct shell file
#>     echo "Creating script bsp2_test_${region}_${feature}"
#>     cat > .${SHORT}.sh <<EOF
#> #!/bin/bash
#> #SBATCH -p shared
#> #SBATCH --mem-per-cpu=10G
#> #SBATCH --job-name=${SHORT}
#> #SBATCH -c 1
#> #SBATCH -o logs/${SHORT}.txt
#> #SBATCH -e logs/${SHORT}.txt
#> #SBATCH --mail-type=ALL
#> 
#> set -e
#> 
#> echo "**** Job starts ****"
#> date
#> 
#> echo "**** JHPCE info ****"
#> echo "User: \${USER}"
#> echo "Job id: \${SLURM_JOB_ID}"
#> echo "Job name: \${SLURM_JOB_NAME}"
#> echo "Node name: \${SLURMD_NODENAME}"
#> echo "Task id: \${SLURM_ARRAY_TASK_ID}"
#> 
#> ## Load the R module
#> module load conda_R
#> 
#> ## List current modules for reproducibility
#> module list
#> 
#> ## Edit with your job command
#> Rscript -e "options(width = 120); print('${region}'); print('${feature}'); sessioninfo::session_info()"
#> 
#> echo "**** Job ends ****"
#> date
#> 
#> ## This script was made using slurmjobs version 0.99.0
#> ## available from http://research.libd.org/slurmjobs/
#> 
#> 
#> EOF
#> 
#>     call="sbatch .${SHORT}.sh"
#>     echo $call
#>     $call
#>     done
#> done
#> 

job_loop(
    loops = list(region = c("DLPFC", "HIPPO"), feature = c("gene", "exon", "tx", "jxn")),
    cores = 5,
    task_num = 10,
    name = "bsp2_test_array"
)
#> #!/bin/bash
#> 
#> ## Usage:
#> # sh bsp2_test_array.sh
#> 
#> ## Create the logs directory
#> mkdir -p logs
#> 
#> for region in DLPFC HIPPO; do
#>     for feature in gene exon tx jxn; do
#> 
#>     ## Internal script name
#>     SHORT="bsp2_test_array_${region}_${feature}"
#> 
#>     # Construct shell file
#>     echo "Creating script bsp2_test_array_${region}_${feature}"
#>     cat > .${SHORT}.sh <<EOF
#> #!/bin/bash
#> #SBATCH -p shared
#> #SBATCH --mem-per-cpu=10G
#> #SBATCH --job-name=${SHORT}
#> #SBATCH -c 5
#> #SBATCH -o logs/${SHORT}.%a.txt
#> #SBATCH -e logs/${SHORT}.%a.txt
#> #SBATCH --mail-type=ALL
#> #SBATCH --array=1-10%20
#> 
#> set -e
#> 
#> echo "**** Job starts ****"
#> date
#> 
#> echo "**** JHPCE info ****"
#> echo "User: \${USER}"
#> echo "Job id: \${SLURM_JOB_ID}"
#> echo "Job name: \${SLURM_JOB_NAME}"
#> echo "Node name: \${SLURMD_NODENAME}"
#> echo "Task id: \${SLURM_ARRAY_TASK_ID}"
#> 
#> ## Load the R module
#> module load conda_R
#> 
#> ## List current modules for reproducibility
#> module list
#> 
#> ## Edit with your job command
#> Rscript -e "options(width = 120); print('${region}'); print('${feature}'); sessioninfo::session_info()"
#> 
#> echo "**** Job ends ****"
#> date
#> 
#> ## This script was made using slurmjobs version 0.99.0
#> ## available from http://research.libd.org/slurmjobs/
#> 
#> 
#> EOF
#> 
#>     call="sbatch .${SHORT}.sh"
#>     echo $call
#>     $call
#>     done
#> done
#>