Build a bash script that loops over variables and submits SLURM jobs

This function builds a bash script functioning as an array job that loops over a set of variables with pre-specified values.

job_loop(
  loops,
  name,
  create_shell = FALSE,
  partition = "shared",
  memory = "10G",
  cores = 1L,
  tc = 20,
  email = "ALL",
  logdir = "logs"
)

Arguments

loops

A named list where each of the elements are character vectors. The names of loops specify the variables used for the loops and the contents specify the options to loop through for each variable.

name

A character(1) vector with the name of the script. Any spaces will be replaced by underscores.

create_shell

A logical(1) vector specifying whether to create a shell file for the script.

partition

A character(1) vector with the name of the SLURM partition. Check how busy a given partition is by running squeue -p [partition].

memory

The amount of memory per core to request. You can check how much a current job is utilizing using the sstat command.

cores

The number of cores to request. Note that the total memory your job will request is cores multiplied by memory.

tc

If task_num is specified, this option controls the number of concurrent tasks.

email

The email reporting option for the email address ("BEGIN", "END", "FAIL", or "ALL")

logdir

The directory for the log files relative to the current working directory.

Value

A character vector with the script contents. If create_shell was specified then it also creates the actual script in the current working directory.

See also

Author

Nicholas J. Eagles

Leonardo Collado-Torres

Examples

job_loop(
    loops = list(region = c("DLPFC", "HIPPO"), feature = c("gene", "exon", "tx", "jxn")),
    name = "bsp2_test_array",
    cores = 2
)
#>  [1] "#!/bin/bash"                                                                                              
#>  [2] "#SBATCH -p shared"                                                                                        
#>  [3] "#SBATCH --mem-per-cpu=10G"                                                                                
#>  [4] "#SBATCH --job-name=bsp2_test_array"                                                                       
#>  [5] "#SBATCH -c 2"                                                                                             
#>  [6] "#SBATCH -o /dev/null"                                                                                     
#>  [7] "#SBATCH -e /dev/null"                                                                                     
#>  [8] "#SBATCH --mail-type=ALL"                                                                                  
#>  [9] "#SBATCH --array=1-8%20"                                                                                   
#> [10] ""                                                                                                         
#> [11] "## Define loops and appropriately subset each variable for the array task ID"                             
#> [12] "all_region=(DLPFC HIPPO)"                                                                                 
#> [13] "region=${all_region[$(( $SLURM_ARRAY_TASK_ID / 4 % 2 ))]}"                                                
#> [14] ""                                                                                                         
#> [15] "all_feature=(gene exon tx jxn)"                                                                           
#> [16] "feature=${all_feature[$(( $SLURM_ARRAY_TASK_ID / 1 % 4 ))]}"                                              
#> [17] ""                                                                                                         
#> [18] "## Explicitly pipe script output to a log"                                                                
#> [19] "log_path=logs/bsp2_test_array_${region}_${feature}_${SLURM_ARRAY_TASK_ID}.txt"                            
#> [20] ""                                                                                                         
#> [21] "{"                                                                                                        
#> [22] "set -e"                                                                                                   
#> [23] ""                                                                                                         
#> [24] "echo \"**** Job starts ****\""                                                                            
#> [25] "date"                                                                                                     
#> [26] ""                                                                                                         
#> [27] "echo \"**** JHPCE info ****\""                                                                            
#> [28] "echo \"User: ${USER}\""                                                                                   
#> [29] "echo \"Job id: ${SLURM_JOB_ID}\""                                                                         
#> [30] "echo \"Job name: ${SLURM_JOB_NAME}\""                                                                     
#> [31] "echo \"Node name: ${SLURMD_NODENAME}\""                                                                   
#> [32] "echo \"Task id: ${SLURM_ARRAY_TASK_ID}\""                                                                 
#> [33] ""                                                                                                         
#> [34] "## Load the R module"                                                                                     
#> [35] "module load conda_R/4.3"                                                                                  
#> [36] ""                                                                                                         
#> [37] "## List current modules for reproducibility"                                                              
#> [38] "module list"                                                                                              
#> [39] ""                                                                                                         
#> [40] "## Edit with your job command"                                                                            
#> [41] "Rscript -e \"options(width = 120); print('${region}'); print('${feature}'); sessioninfo::session_info()\""
#> [42] ""                                                                                                         
#> [43] "echo \"**** Job ends ****\""                                                                              
#> [44] "date"                                                                                                     
#> [45] ""                                                                                                         
#> [46] "} > $log_path 2>&1"                                                                                       
#> [47] ""                                                                                                         
#> [48] "## This script was made using slurmjobs version 0.99.0"                                                   
#> [49] "## available from http://research.libd.org/slurmjobs/"                                                    
#> [50] ""